Combine the sleep, retry, and timeout customizations
