+++ codegen/src/main/kotlin/software/amazon/smithy/rust/codegen/smithy/customize/RequiredCustomizations.kt
@@ -16,11 +16,9 @@ import software.amazon.smithy.rust.codegen.smithy.customizations.EndpointPrefixG
 import software.amazon.smithy.rust.codegen.smithy.customizations.HttpChecksumRequiredGenerator
 import software.amazon.smithy.rust.codegen.smithy.customizations.HttpVersionListCustomization
 import software.amazon.smithy.rust.codegen.smithy.customizations.IdempotencyTokenGenerator
-import software.amazon.smithy.rust.codegen.smithy.customizations.PubUseRetryConfigGenerator
-import software.amazon.smithy.rust.codegen.smithy.customizations.RetryConfigProviderCustomization
-import software.amazon.smithy.rust.codegen.smithy.customizations.SleepImplProviderCustomization
+import software.amazon.smithy.rust.codegen.smithy.customizations.ResiliencyConfigCustomization
+import software.amazon.smithy.rust.codegen.smithy.customizations.ResiliencyReExportCustomization
 import software.amazon.smithy.rust.codegen.smithy.customizations.SmithyTypesPubUseGenerator
-import software.amazon.smithy.rust.codegen.smithy.customizations.TimeoutConfigProviderCustomization
 import software.amazon.smithy.rust.codegen.smithy.generators.LibRsCustomization
 import software.amazon.smithy.rust.codegen.smithy.generators.config.ConfigCustomization
 
@@ -48,10 +46,7 @@ class RequiredCustomizations : RustCodegenDecorator<ClientCodegenContext> {
         codegenContext: ClientCodegenContext,
         baseCustomizations: List<ConfigCustomization>,
     ): List<ConfigCustomization> =
-        baseCustomizations +
-            RetryConfigProviderCustomization(codegenContext) +
-            SleepImplProviderCustomization(codegenContext) +
-            TimeoutConfigProviderCustomization(codegenContext)
+        baseCustomizations + ResiliencyConfigCustomization(codegenContext)
 
     override fun libRsCustomizations(
         codegenContext: ClientCodegenContext,
@@ -60,7 +55,7 @@ class RequiredCustomizations : RustCodegenDecorator<ClientCodegenContext> {
         baseCustomizations + CrateVersionGenerator() +
             SmithyTypesPubUseGenerator(codegenContext.runtimeConfig) +
             AllowLintsGenerator() +
-            PubUseRetryConfigGenerator(codegenContext.runtimeConfig)
+            ResiliencyReExportCustomization(codegenContext.runtimeConfig)
 
     override fun extras(codegenContext: ClientCodegenContext, rustCrate: RustCrate) {
         // Add rt-tokio feature for `ByteStream::from_path`
