+++ codegen/src/test/kotlin/software/amazon/smithy/rust/codegen/customizations/ResiliencyConfigCustomizationTest.kt
@@ -6,7 +6,7 @@
 package software.amazon.smithy.rust.codegen.customizations
 
 import org.junit.jupiter.api.Test
-import software.amazon.smithy.rust.codegen.smithy.customizations.TimeoutConfigProviderCustomization
+import software.amazon.smithy.rust.codegen.smithy.customizations.ResiliencyConfigCustomization
 import software.amazon.smithy.rust.codegen.smithy.transformers.OperationNormalizer
 import software.amazon.smithy.rust.codegen.smithy.transformers.RecursiveShapeBoxer
 import software.amazon.smithy.rust.codegen.testutil.TestWorkspace
@@ -15,7 +15,7 @@ import software.amazon.smithy.rust.codegen.testutil.rustSettings
 import software.amazon.smithy.rust.codegen.testutil.testCodegenContext
 import software.amazon.smithy.rust.codegen.testutil.validateConfigCustomizations
 
-internal class TimeoutConfigDecoratorTest {
+internal class ResiliencyConfigCustomizationTest {
     private val baseModel = """
         namespace test
         use aws.protocols#awsQuery
@@ -38,6 +38,6 @@ internal class TimeoutConfigDecoratorTest {
         val project = TestWorkspace.testProject()
         val codegenContext = testCodegenContext(model, settings = project.rustSettings())
 
-        validateConfigCustomizations(TimeoutConfigProviderCustomization(codegenContext), project)
+        validateConfigCustomizations(ResiliencyConfigCustomization(codegenContext), project)
     }
 }
